requirements.txt
alembic==1.13.2
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.2.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
Mako==1.3.5
MarkupSafe==2.1.5
numpy==1.26.4
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pydantic_core==2.18.4
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.10.0
uvicorn==0.29.0
watchfiles==0.22.0
websockets==12.0
